fix: avoid undefined in tailwind config and rename v2 styled demo
